requirements.txt
streamlit==1.39.0
python-dotenv==1.0.1
Pillow==10.4.0
numpy==2.1.3
opencv-python-headless==4.10.0.84
easyocr==1.7.1
torch>=2.2,<3
groq==0.11.0
requests==2.32.3
